Start stripping
