+++ cabal.project
@@ -2,11 +2,11 @@ packages:
   .
   ../forks/http2
 
---  ~/wt/ghc/repo/9.2.8-patched/patched-libs/async
---  ~/wt/ghc/repo/9.2.8-patched/patched-libs/tasty/core
---  ~/wt/ghc/repo/9.2.8-patched/patched-libs/unliftio/unliftio-core
---  ~/wt/ghc/repo/9.2.8-patched/patched-libs/unliftio/unliftio
---  ~/wt/ghc/repo/9.2.8-patched/patched-libs/lens
+  ~/wt/ghc/repo/9.2.8-patched/patched-libs/async
+  ~/wt/ghc/repo/9.2.8-patched/patched-libs/tasty/core
+  ~/wt/ghc/repo/9.2.8-patched/patched-libs/unliftio/unliftio-core
+  ~/wt/ghc/repo/9.2.8-patched/patched-libs/unliftio/unliftio
+  ~/wt/ghc/repo/9.2.8-patched/patched-libs/lens
 
 --   ~/wt/packages/stm-debug
 --   ~/wt/packages/stm-debug/assignid
@@ -18,12 +18,10 @@ package grapesy
   tests: True
   flags: +build-demo +build-stress-test
 
--- The version on Hackage also works, it just writes some (unnecessary)
--- exceptions to stderr
--- source-repository-package
---   type: git
---   location: https://github.com/edsko/http2.git
---   tag: c2a2994ed45a08998c2b6eb22a08f28b9eb36c3c
-
 allow-newer: stm
 
+-- https://github.com/kazu-yamamoto/http2-tls/pull/4
+source-repository-package
+  type: git
+  location: https://github.com/edsko/http2-tls.git
+  tag: 4fdfe7c32177c5a483916623cb8f93525761cff6
