+++ grapesy.cabal
@@ -214,10 +214,6 @@ test-suite test-grapesy
       Test.Driver.Dialogue.Execution
       Test.Driver.Dialogue.Generation
       Test.Driver.Dialogue.TestClock
-      Test.Prop.Dialogue
-      Test.Prop.Serialization
-      Test.Sanity.HalfClosedLocal
-      Test.Sanity.StreamingType.NonStreaming
       Test.Util.ClientServer
       Test.Util.PrettyVal
   build-depends:
@@ -242,132 +238,6 @@ test-suite test-grapesy
     , text             >= 1.2  && < 2.1
     , tls              >= 1.5  && < 1.8
 
-executable demo-client
-  import:
-      lang
-  hs-source-dirs:
-      demo-client
-      demo-common
-      proto
-  main-is:
-      Main.hs
-  ghc-options:
-      -threaded
-  other-modules:
-      Demo.Client.API.Core.Greeter
-      Demo.Client.API.Core.NoFinal.Greeter
-      Demo.Client.API.Core.RouteGuide
-      Demo.Client.API.Protobuf.Greeter
-      Demo.Client.API.Protobuf.Pipes.RouteGuide
-      Demo.Client.API.Protobuf.RouteGuide
-      Demo.Client.Cmdline
-      Demo.Client.Util.DelayOr
-      Demo.Common.Logging
-      Proto.Helloworld
-      Proto.Helloworld_Fields
-      Proto.RouteGuide
-      Proto.RouteGuide_Fields
-  build-depends:
-      -- Internal dependencies
-    , grapesy
-  build-depends:
-      -- External dependencies
-    , async                >= 2.2  && < 2.3
-    , contra-tracer        >= 0.2  && < 0.3
-    , data-default         >= 0.7  && < 0.8
-    , lens                 >= 5.0  && < 5.3
-    , optparse-applicative >= 0.16 && < 0.19
-    , pipes                >= 4.3  && < 4.4
-    , pipes-safe           >= 2.3  && < 2.4
-    , proto-lens           >= 0.7  && < 0.8
-    , proto-lens-runtime   >= 0.7  && < 0.8
-    , stm                  >= 2.5  && < 2.6
-    , text                 >= 1.2  && < 2.1
-
-  if !flag(build-demo)
-    buildable:
-      False
-
-executable demo-server
-  import:
-      lang
-  hs-source-dirs:
-      demo-server
-      demo-common
-      proto
-  main-is:
-      Main.hs
-  ghc-options:
-      -threaded
-  other-modules:
-      Demo.Common.Logging
-      Demo.Server.Aux.RouteGuide
-      Demo.Server.Cmdline
-      Demo.Server.Service.Greeter
-      Demo.Server.Service.RouteGuide
-      Proto.Helloworld
-      Proto.Helloworld_Fields
-      Proto.RouteGuide
-      Proto.RouteGuide_Fields
-      Paths_grapesy
-  build-depends:
-      -- Internal dependencies
-    , grapesy
-  build-depends:
-      -- External dependencies
-    , aeson                >= 1.5  && < 2.2
-    , containers           >= 0.6  && < 0.7
-    , contra-tracer        >= 0.2  && < 0.3
-    , data-default         >= 0.7  && < 0.8
-    , lens                 >= 5.0  && < 5.3
-    , mtl                  >= 2.2  && < 2.4
-    , network              >= 3.1  && < 3.2
-    , optparse-applicative >= 0.16 && < 0.19
-    , proto-lens           >= 0.7  && < 0.8
-    , proto-lens-runtime   >= 0.7  && < 0.8
-    , text                 >= 1.2  && < 2.1
-    , time                 >= 1.9  && < 1.13
-    , transformers         >= 0.5  && < 0.7
-
-  if !flag(build-demo)
-    buildable:
-      False
-
-executable test-stress
-  import:
-      lang
-  hs-source-dirs:
-      test-stress
-      test-common
-  main-is:
-      Main.hs
-  ghc-options:
-      -threaded
-  other-modules:
-      Test.Stress.Cmdline
-      Test.Util.ClientServer
-      Test.Util.PrettyVal
-      Paths_grapesy
-  build-depends:
-      -- Internal dependencies
-    , grapesy
-  build-depends:
-      -- External dependencies
-    , async                >= 2.2  && < 2.3
-    , containers           >= 0.6  && < 0.7
-    , contra-tracer        >= 0.2  && < 0.3
-    , data-default         >= 0.7  && < 0.8
-    , exceptions           >= 0.10 && < 0.11
-    , http2
-    , optparse-applicative >= 0.16 && < 0.19
-    , pretty-show          >= 1.10 && < 1.11
-    , stm                  >= 2.5  && < 2.6
-    , text                 >= 1.2  && < 2.1
-    , tls                  >= 1.5  && < 1.8
-
-  if !flag(build-stress-test)
-    buildable:
-      False
 
 Flag build-demo
   description: Build the demo
